tests/wpt/tests/css/css-box/margin-trim/computed-margin-values/flexbox-row-multi-line-inline-end.html

<!DOCTYPE html>
<html>
<head>
<link rel="author" href="mailto:sammy.gill@apple.com">
<link rel="help" href="https://drafts.csswg.org/css-box-4/#margin-trim-flex">
<meta name="assert" content="trimmed inline-end margins should be reflected in computed style">
<style>
flexbox {
    display: flex;
    width: 110px;
    flex-wrap: wrap;
    margin-trim: inline-end;
}
item {
    display: block;
    background-color: green;
    width: 50px;
    height: 50px;
}
item:nth-child(1) {
    margin-inline-end: 10px;
}
item:nth-child(2) {
    margin-inline-end: -10px;
}
item:nth-child(3) {
    margin-inline-end: 10px;
}
item:nth-child(4) {
    margin-inline-end: 50%;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
</head>
<body onload="checkLayout('flexbox > item')">
<div id="target">
<flexbox>
    <item data-expected-margin-right="10" data-offset-x="8"></item>
    <item data-expected-margin-right="0" data-offset-x="68"></item>
    <item data-expected-margin-right="10" data-offset-x="8"></item>
    <item data-expected-margin-right="0" data-offset-x="68"></item>
</flexbox>
</div>
</body>
</html>
